changed types and tested dataset with file path

+++ hypex/utils/hypex_typings.py
@@ -1,9 +1,7 @@
 import datetime
 from typing import Any, Union
 
-import numpy as np
-
-StratificationRoleTypes = Union[np.float64, str, datetime.datetime]
-TreatmentRoleTypes = Union[np.float64, bool, str]
-TargetRoleTypes = Union[np.float64, int, bool]
+StratificationRoleTypes = Union[float, str, datetime.datetime]
+TreatmentRoleTypes = Union[float, bool, str]
+TargetRoleTypes = Union[float, int, bool]
 ABCRoleTypes = Union[Any]
